0.5.22 — Servitor terminal reconnect fix + audit batch (12 real bugs) + monitor self-heal logging
Servitor: appliance restore-on-load now dispatches a BUBBLING change event, so the
document-delegated handler runs and the terminal reconnects (SSH) / hides (repo) on
reload — previously the value restored but the terminal side-effects never fired.

Audit fixes (each verified against source; the two "high severity" findings —
repo-worker prompt swap, tool-name index mismatch — were FALSE positives, skipped):
- core/vector_store.go: maintenance funcs keyed by Key (not Label); SplitReportIntoChunks
  computes the paragraph split once; SearchChunksSubstring joins section/text with "\n",
  carries Kind, and uses Score:0 to match the predicate variants; hard-cut is rune-safe
  (utf8.RuneStart backup) so a chunk boundary never splits a multi-byte character
- core/llm_openai.go: estimateReasoningTokens uses an int64 product (no 32-bit overflow)
- apps/bridges: upsertMember is case-insensitive (no duplicate members); handlePanic
  preserves SelfName/SelfHandle instead of wiping the config; keys PATCH returns 400 on a
  bad body; group self-heal re-stamps ALL stale bindings (not just the first); the dispatch
  goroutine is bounded by a 10-minute timeout so a hung LLM/tool call can't leak it
- apps/servitor: repo WalkDir logs skipped entries; the watch loop persists NextRunAt
  BEFORE launching the check goroutine, so a completed watch's Done=true isn't clobbered
  back to false (no spurious re-fire)

Also: event-monitor startup self-heal logs how many monitors it re-armed (observability).

+++ core/llm_openai.go
@@ -1702,7 +1702,9 @@ func estimateReasoningTokens(total int, reasoning, content string, toolCalls []T
 	if denom == 0 {
 		return 0
 	}
-	return total * rChars / denom
+	// int64 for the product: total*rChars can exceed 2^31 (e.g. 32k tokens *
+	// 100k chars) and overflow a 32-bit int before the divide.
+	return int(int64(total) * int64(rChars) / int64(denom))
 }
 
 // ChatStream sends a streaming request.
